Adding rudimentary cstdlib for user programs

+++ src/kernel/sched/stack_check.c
@@ -3,7 +3,7 @@
 void stack_compute_checksum(STACK_CHECKSUM* checksum, const void *esp, const void *ebp) {
     *checksum = 0;
     
-    for (uint32_t *p = esp; p < ebp; p = &p[1]) {
+    for (uint32_t *p = (uint32_t)(esp); p < (uint32_t)(ebp); p = &p[1]) {
         *checksum ^= *p;
     }
 }
